Add files via upload
Clustering: It is organizing data points into clusters based on Euclidean distance. we have 3 types of clustering 1. K-Means Clusters , 2. DB Scan and 3. Agglomerative Clustering.
Mostly we will use K-Means Clustering. in a dataset how many clusters we have to create depends upon Elbow method. In Elbow method X-Axis we have no of cluster and on Y-Axis we will use WCSS(with in clusters sum of squares). In clustering part we dont have dependent variable and we are not required any scaling technique.